+++ test/VaultHelper.sol
@@ -13,7 +13,7 @@ contract VaultHelper is Vault {
 
     function totalSupply(uint256 epochId) public view returns (uint256) {
         ERC7540Storage storage $ = _getERC7540Storage();
-        return $.settles[$.epochs[epochId].settleId].totalSupply;
+        return $.settles[$.epochs[uint40(epochId)].settleId].totalSupply;
     }
 
     // function previousEpochTotalSupply() public view returns (uint256) {
@@ -23,7 +23,7 @@ contract VaultHelper is Vault {
 
     function totalAssets(uint256 epochId) public view returns (uint256) {
         ERC7540Storage storage $ = _getERC7540Storage();
-        return $.settles[$.epochs[epochId].settleId].totalAssets;
+        return $.settles[$.epochs[uint40(epochId)].settleId].totalAssets;
     }
 
     function newTotalAssets() public view returns (uint256) {
